django_oz/pyproject.toml
[project]
name = "django-oz"
version = "0.1.0"
description = ""
authors = [
    {name = "yoo-sh-96",email = "tmdguq90051@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
